+++ app/src/main/java/com/dms/pmsandroid/feature/introduce/model/ClubDetailModel.kt
@@ -1,0 +1,4 @@
+package com.dms.pmsandroid.feature.introduce.model
+
+class ClubDetailModel {
+}
